implement standalone cli switch class
