+++ source/main.ts
@@ -14,14 +14,18 @@
  */
 
 import { app } from 'electron'
-import path from 'path'
 import { bootApplication, shutdownApplication } from './app/lifecycle'
+
+// Helper function to extract files to open from process.argv
+import extractFilesFromArgv from './app/util/extract-files-from-argv'
 
 // Include the global Zettlr class
 import Zettlr from './main/zettlr'
 
-// Helper function to extract files to open from process.argv
-import extractFilesFromArgv from './app/util/extract-files-from-argv'
+// Include the global Zettlr class
+import Commandlineswitches from './commandlineswitches'
+const c = new Commandlineswitches()
+c.handleCLI()
 
 // Immediately after launch, check if there is already another instance of
 // Zettlr running, and, if so, exit immediately. The arguments (including files)
@@ -62,38 +66,6 @@ global.log = {
 // See https://www.electronjs.org/docs/tutorial/notifications#windows
 if (process.platform === 'win32') {
   app.setAppUserModelId('com.zettlr.app')
-}
-
-// Setting custom data dir for user configuration files.
-// Full path or relative path is OK. '~' does not work as expected.
-const dataDirFlag = process.argv.find(elem => elem.indexOf('--data-dir=') === 0)
-
-if (dataDirFlag !== undefined) {
-  // a path to a custom config dir is provided
-  const match = /^--data-dir="?([^"]+)"?$/.exec(dataDirFlag)
-  if (match !== null) {
-    let dataDir = match[1]
-
-    if (!path.isAbsolute(dataDir)) {
-      if (app.isPackaged) {
-        // Attempt to use the executable file's path as the basis
-        dataDir = path.join(path.dirname(app.getPath('exe')), dataDir)
-      } else {
-        // Attempt to use the repository's root directory as the basis
-        dataDir = path.join(__dirname, '../../', dataDir)
-      }
-    }
-    global.log.info('[Application] Using custom data dir: ' + dataDir)
-    app.setPath('userData', dataDir)
-    app.setAppLogsPath(path.join(dataDir, 'logs'))
-  }
-}
-
-// On systems with virtual GPUs (i.e. VMs), it might be necessary to disable
-// hardware acceleration. If the corresponding flag is set, we do so.
-// See for more info https://github.com/Zettlr/Zettlr/issues/2127
-if (process.argv.includes('--disable-hardware-acceleration')) {
-  app.disableHardwareAcceleration()
 }
 
 // *****************************************************************************
@@ -149,6 +121,10 @@ app.whenReady().then(() => {
         // After the app has been booted, open any files that we amassed in the
         // meantime.
         zettlr?.handleAddRoots(filesBeforeOpen)
+        // transfer the cli switches to the main config so that it can be used later
+        for (let [ key, value ] of cli.getSwitches().entries()) {
+          global.config.set(key, value)
+        }
       })
       .catch(err => {
         console.error(err)
